requirements.txt
httpx
python-dotenv
google-genai
edge-tts
nest_asyncio
